Move Parser::pipeline
